[refactor] 기능 함수 분리로 코드 구조개선 기능함수들을 funcs 폴더에 추가해 코드 구조개선
